Move stack upgrade test into integration test for parallelism
